# functions/openapi.yaml
openapi: 3.1.0
info:
  title: LumiMD API
  version: 1.0.0
  description: |
    Backend API for LumiMD - a patient-facing medical navigation app.
    
    ## Authentication
    Most endpoints require a Firebase ID token in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```
    
    ## Error Format
    All errors follow this format:
    ```json
    {
      "code": "error_code",
      "message": "Human readable message",
      "details": { "field": "optional context" }
    }
    ```
servers:
  - url: https://us-central1-lumimd.cloudfunctions.net/api
    description: Production
  - url: https://us-central1-lumimd-dev.cloudfunctions.net/api
    description: Development
...
paths:
...
  /v1/meds:
    get:
      tags: [meds]
      summary: List medications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of medications
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Medication'

    post:
      tags: [meds]
      summary: Create medication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationInput'
      responses:
        '201':
          description: Medication created
  /v1/meds/{id}:
    patch:
      tags: [meds]
      summary: Update medication
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationInput'
      responses:
        '200':
          description: Medication updated

    delete:
      tags: [meds]
      summary: Delete medication
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Medication deleted

  # ==================== Sharing ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  schemas:
...
    Medication:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        dose:
          type: string
        frequency:
          type: string
        notes:
          type: string
    MedicationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Lisinopril"
        dose:
          type: string
          example: "10 mg"
        frequency:
          type: string
          example: "Daily"
        notes:
          type: string
...